requirements.txt
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
scipy==1.10.1
scikit-learn==1.2.2
python-dotenv